backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==43.0.1
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==4.2.16
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-base64==2.0
fonttools==4.54.1
fpdf==1.7.2
gunicorn==23.0.0
html5lib==1.1
idna==3.10
inflection==0.5.1
isort==5.13.2
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.16.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
webencodings==0.5.1
zopfli==0.2.3
django-cors-headers==4.6.0
drf-extra-fields==3.7.0